♻️ factor out choice generation
